feat(rsc): split server and client runtimes
