+++ packages/plugin-rsc/src/react/rsc.ts
@@ -1,90 +1,2 @@
-// @ts-ignore
-import * as ReactClient from '@vitejs/plugin-rsc/vendor/react-server-dom/client.edge'
-// @ts-ignore
-import * as ReactServer from '@vitejs/plugin-rsc/vendor/react-server-dom/server.edge'
-import type { ReactFormState } from 'react-dom/client'
-import {
-  createClientManifest,
-  createServerDecodeClientManifest,
-  createServerManifest,
-} from '../core/rsc'
-import type {
-  ClientTemporaryReferenceSet,
-  CreateFromReadableStreamEdgeOptions,
-  DecodeReplyFunction,
-  EncodeReplyFunction,
-  RenderToReadableStreamOptions,
-  ServerTemporaryReferenceSet,
-} from '../types'
-
-export { loadServerAction, setRequireModule } from '../core/rsc'
-
-export function renderToReadableStream<T>(
-  data: T,
-  options?: RenderToReadableStreamOptions,
-  extraOptions?: {
-    /**
-     * @internal
-     */
-    onClientReference?: (metadata: { id: string; name: string }) => void
-  },
-): ReadableStream<Uint8Array> {
-  return ReactServer.renderToReadableStream(
-    data,
-    createClientManifest({
-      onClientReference: extraOptions?.onClientReference,
-    }),
-    options,
-  )
-}
-
-export function createFromReadableStream<T>(
-  stream: ReadableStream<Uint8Array>,
-  options: CreateFromReadableStreamEdgeOptions = {},
-): Promise<T> {
-  return ReactClient.createFromReadableStream(stream, {
-    serverConsumerManifest: {
-      // https://github.com/facebook/react/pull/31300
-      // https://github.com/vercel/next.js/pull/71527
-      serverModuleMap: createServerManifest(),
-      moduleMap: createServerDecodeClientManifest(),
-    },
-    ...options,
-  })
-}
-
-export function registerClientReference<T>(
-  proxy: T,
-  id: string,
-  name: string,
-): T {
-  return ReactServer.registerClientReference(proxy, id, name)
-}
-
-export const registerServerReference: <T>(
-  ref: T,
-  id: string,
-  name: string,
-) => T = ReactServer.registerServerReference
-
-export const decodeReply: DecodeReplyFunction = (body, options) =>
-  ReactServer.decodeReply(body, createServerManifest(), options)
-
-export function decodeAction(body: FormData): Promise<() => Promise<void>> {
-  return ReactServer.decodeAction(body, createServerManifest())
-}
-
-export function decodeFormState(
-  actionResult: unknown,
-  body: FormData,
-): Promise<ReactFormState | undefined> {
-  return ReactServer.decodeFormState(actionResult, body, createServerManifest())
-}
-
-export const createTemporaryReferenceSet: () => ServerTemporaryReferenceSet =
-  ReactServer.createTemporaryReferenceSet
-
-export const encodeReply: EncodeReplyFunction = ReactClient.encodeReply
-
-export const createClientTemporaryReferenceSet: () => ClientTemporaryReferenceSet =
-  ReactClient.createTemporaryReferenceSet
+export * from './rsc/server'
+export * from './rsc/client'
