+++ packages/plugin-rsc/src/rsc/client.ts
@@ -1,0 +1,3 @@
+import './shared'
+
+export * from '../react/rsc/client'
