+++ packages/plugin-rsc/tsdown.config.ts
@@ -8,6 +8,8 @@ export default defineConfig({
     'src/browser.ts',
     'src/ssr.tsx',
     'src/rsc.tsx',
+    'src/rsc/server.tsx',
+    'src/rsc/client.ts',
     'src/core/browser.ts',
     'src/core/ssr.ts',
     'src/core/rsc.ts',
@@ -15,6 +17,8 @@ export default defineConfig({
     'src/react/browser.ts',
     'src/react/ssr.ts',
     'src/react/rsc.ts',
+    'src/react/rsc/server.ts',
+    'src/react/rsc/client.ts',
     'src/transforms/index.ts',
     'src/plugins/cjs.ts',
     'src/utils/rpc.ts',
